Fix parsing mock GPS data from Serial

+++ src/DataPoint.h
@@ -32,7 +32,7 @@
 class DataPoint
 {
 public:
-    long        ts;
+    long long        ts;
 
     bool        hasGeodetic;
     bool        isValid;
